+++ beyond/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2 
+
+- Fix env path
+- Change DB insert, to DB create
+- Update DB for multiple insert records
+
 ## 0.0.1
 
 - Initial upload for alpha testing